[android][audio] Add base class for players
